requirements.txt
Flask
Flask-Session
flaskWebGUI
web3
py-solc-x
python-dotenv
requests

